filling out comments, get-intunehardwareids works w/o wrapper now
i removed the need for wrapper function for get-windowsautopilotinfo, when called through get-intunehardwareids
The script tries to install get-windowsautopilotinfo from the internet if its not found in the suppoortfiles directory.
I read in a book that using a splat to define parameters gives you the advantage of only having one set to modify (in situations where you'd otherwise write the same set of parameters out explicitly, two+ times).