Pipeline / parameter input working and tested on most functions
Editing most report generating functions to get pipeline AND parameter input for TargetComputer to work correctly with begin/process/end structure.

Adding foreach loop in the process block helps to account for both methods of input. because TargetComputer is only a list type structure in the process block if its been passed in through the begin block.

Also adding some lines during output file handling in the begin blocks to create any directories that need to be created when the functions aren't being run as a part of the terminal menu.

Adding a try/catch to the end of these functions, that tries to open the report directory (if reports were generated) using the parent property of the outputfile variable. If something goes wrong trying to open the directory (for ex if its null), the script defaults to opening the generated .csv file.

*Check values / calculations in ping test report. - packet loss percentage being reported as 100 for offline device

UNTESTED: Get-USBDevices, Install-NewTeams/SmartNotebook/VeyonRoom, scanoffice2021, scan-softwareinventory, stop-wifiadapters, send-reboots, send-files