Adding List Dot Net versions function
Function is able to be executed through menu and presents a gridview of .NET installations on each target computer in the list. I was able to create this function by pasting in the repeatable code sections:
Begin block - Targetcomputer processing
PRocess block - Testing each TargetComputer to see if it's responsive on network, then running the scriptblock (in this case - creating an object for each dotnet version found on each computer).
Outputs to Gridview, and I'm hoping to eventually just have one row for each computer, listing all dotnet versions to make it a bit easier to read.