Testing 'job' idea
created completedjobs directory
created Install-ApplicationJob function, which is a copy of Install-Application, except the AppName parameter is mandatory, along with a few other changes so that the function would not require any user input (addition of do not disturb to function parameters being one).

Menu.ps1 reads the config.json file to determine which functions should be run as jobs, and creates a file system watcher object to watch the completedjobs directory for new files.

When a job is completed - it will output an announcement with any relavant info to a text file in ./completedjobs.

Unsure how to navigate errors, timeouts, etc. in the functions being run as jobs at this point, but install-applicationjob is ready for it's first test runs.

'Separatejobfunctions.txt' in base directory contains functions I think would be good to run as background jobs. In most cases, these are functions that 1. take a bit longer, and 2. can be adapted so they don't require user input.