added watcher.exe
Right now, watcher.ps1 is working correctly and opening any text files created in completedjobs directory while the menu is running.
The issue is that the file system watcher continues to operate even after you've exited/closed the terminal menu. The filesystem watcher is tied to another powershell process, so I'm looking into ways of titling the process or id'ing it somehow, so it can be closed automatically when the menu closes.