Jobs testing successfully (so far)
Several tests run successfully using a new structure for part of menu.ps1.

I only changed a section to make sure I could get it to work - the section that runs functions that have TargetComputer, and pipes that value into the command itself.

Instead of making you wait until the function has completed to allow you to go back to the menu and run a new function - the Start-Job cmdlet is used to run the command in the background. Passing the arguments in was tricky, but some of the articles listed here: https://duckduckgo.com/?q=start+powershell+in+a+new+window&ia=web were a big help.