Readme.md, menu.ps1 changes
Changed executionpolicy command in readme to set to unrestricted instead of bypass. Hoping this will prevent some of the prompts that occur for new computers/users.

added section to menu.ps1 that checks if dealing with a parameter named outputfile when prompting the user for values - if so, it will append the current working directory's absolute path to the output file value.

Added 'import-module importexcel' to get-currentuser, get-assetinformation, and get-computerdetails for testing - I think .xlsx file creation is getting tripped up when running the functions in background.