Fixing functions
Copy snippet working / tested.

Copy canned response improved by replacing both  (($ and $))  with '' (nothing), instead of just replacing (($.
Before, variables defined in canned response files that didn't have descriptions in them would be displayed in the terminal with the ending $)).

Also added section that checks if there is a description and prints it to terminal when prompting the user for a value if so.

New-DriveMappingExe tested