Update Add-PrinterLogicPrinter.ps1
If there are no results from function (if all target computers were offline) function will still create a text file and attempt to open it. this is for when the function is run as a job in the background, so the user still gets a notification that it worked and its complete.